Issue #SB-10943 Fix: Content enriched metadata was not getting synced as part of hierarchy update
