+++ platform-jobs/samza/distribution/pom.xml
@@ -53,7 +53,7 @@
 		<dependency>
 			<groupId>org.ekstep</groupId>
 			<artifactId>publish-pipeline</artifactId>
-			<version>0.0.212</version>
+			<version>0.0.213</version>
 			<type>tar.gz</type>
 			<classifier>distribution</classifier>
 		</dependency>
