+++ platform-jobs/samza/publish-pipeline/pom.xml
@@ -7,7 +7,7 @@
 		<version>1.0-SNAPSHOT</version>
 	</parent>
 	<artifactId>publish-pipeline</artifactId>
-	<version>0.0.212</version>
+	<version>0.0.213</version>
 	<dependencies>
 		<dependency>
 			<groupId>org.ekstep</groupId>
