+++ platform-modules/actors/src/main/java/org/ekstep/learning/util/ControllerUtil.java
@@ -236,7 +236,7 @@ public class ControllerUtil extends BaseLearningManager {
 	}
 
 	public Response getHirerachy(String identifier) {
-		String url = Platform.config.getString("platform-api-url") + "/content/v3/hierarchy/" + identifier;
+		String url = Platform.config.getString("platform-api-url") + "/content/v3/hierarchy/" + identifier + "?mode=edit";
 		Response hirerachyRes = null;
 		try {
 			String result = HTTPUtil.makeGetRequest(url);
