+++ platform-modules/content-manager/src/main/java/org/ekstep/content/operation/finalizer/PublishFinalizer.java
@@ -1,6 +1,15 @@
 package org.ekstep.content.operation.finalizer;
 
-import com.rits.cloning.Cloner;
+import java.io.File;
+import java.net.URL;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
+
 import org.apache.commons.lang3.BooleanUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.ekstep.common.Platform;
@@ -23,22 +32,11 @@ import org.ekstep.content.util.ContentPackageExtractionUtil;
 import org.ekstep.graph.dac.enums.GraphDACParams;
 import org.ekstep.graph.dac.model.Node;
 import org.ekstep.graph.engine.router.GraphEngineManagers;
-import org.ekstep.graph.model.node.DefinitionDTO;
 import org.ekstep.graph.service.common.DACConfigurationConstants;
 import org.ekstep.learning.contentstore.VideoStreamingJobRequest;
-import org.ekstep.learning.hierarchy.store.HierarchyStore;
-import org.ekstep.learning.util.ControllerUtil;
 import org.ekstep.telemetry.logger.TelemetryManager;
 
-import java.io.File;
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
+import com.rits.cloning.Cloner;
 
 /**
  * The Class BundleFinalizer, extends BaseFinalizer which mainly holds common
@@ -67,11 +65,6 @@ public class PublishFinalizer extends BaseFinalizer {
 	private static final String ECML_MIMETYPE = "application/vnd.ekstep.ecml-archive";
 	
 	private static ContentPackageExtractionUtil contentPackageExtractionUtil = new ContentPackageExtractionUtil();
-
-	private ControllerUtil util = new ControllerUtil();
-
-	private HierarchyStore hierarchyStore = new HierarchyStore();
-
 
 	/**
 	 * Instantiates a new PublishFinalizer and sets the base path and current
@@ -381,24 +374,10 @@ public class PublishFinalizer extends BaseFinalizer {
 					String.valueOf(node.getMetadata().get(ContentWorkflowPipelineParams.pkgVersion.name())));
 		}
 		
-		if (StringUtils.equalsIgnoreCase(
-				((String) newNode.getMetadata().get(ContentWorkflowPipelineParams.mimeType.name())),
-				COLLECTION_MIMETYPE)) {
-			Node publishedNode = util.getNode(TAXONOMY_ID, contentId);
-			publishHierarchy(publishedNode);
-		}
-
 		TelemetryManager.log("Generating Telemetry Event. | [Content ID: " + contentId + "]");
 		newNode.getMetadata().put(ContentWorkflowPipelineParams.prevState.name(),
 				ContentWorkflowPipelineParams.Processing.name());
 		return response;
-	}
-
-
-	private void publishHierarchy(Node publishedNode) {
-		DefinitionDTO definition = util.getDefinition(publishedNode.getGraphId(), publishedNode.getObjectType());
-		Map<String, Object> hierarchy = util.getHierarchyMap(publishedNode.getGraphId(), publishedNode.getIdentifier(), definition, null, null);
-		hierarchyStore.saveOrUpdateHierarchy(publishedNode.getIdentifier(), hierarchy);
 	}
 
 	/**
